Bloquear escritura

+++ src/AppBundle/Controller/Distribution_companyController.php
@@ -22,6 +22,15 @@ class Distribution_companyController extends Controller
      */
     public function newDistributionCompanyAction(Request $request)
     {
+        $current_convocatory = $this->getUser()->getCurrentConvocatory();
+        if(!$this->get('app.functionsHelper')->isConvocatoryValid($current_convocatory)) {
+            $request->getSession()
+                ->getFlashBag()
+                ->add('error', 'Convocatoria antigua (Solo lectura)')
+            ;
+            return $this->redirectToRoute('user_fct');
+        }
+
         $redirect = 'user_fct';
 
         $distribution = new Distribution_company();
@@ -75,6 +84,19 @@ class Distribution_companyController extends Controller
     public function editCompanyAction(Request $request, Distribution_company $distribution)
     {
         $redirect = 'user_fct';
+        if ($request->get('flag') == 'index') {
+            $redirect = 'index_web';
+        }
+
+        $current_convocatory = $this->getUser()->getCurrentConvocatory();
+        if(!$this->get('app.functionsHelper')->isConvocatoryValid($current_convocatory)) {
+            $request->getSession()
+                ->getFlashBag()
+                ->add('error', 'Convocatoria antigua (Solo lectura)')
+            ;
+            return $this->redirectToRoute($redirect);
+        }
+
         $current_user = $this->getUser();
 
         $options = Array(
@@ -111,10 +133,6 @@ class Distribution_companyController extends Controller
 
         }
 
-        if ($request->get('flag') == 'index') {
-            $redirect = 'index_web';
-        }
-
         return $this->render('user/forms/form.html.twig', array(
             'form' => $form->createView(),
             'title' => "Modificar asignación FCT",
@@ -127,6 +145,15 @@ class Distribution_companyController extends Controller
      */
     public function deleteCompanyAction(Request $request, Distribution_company $distribution_company)
     {
+        $current_convocatory = $this->getUser()->getCurrentConvocatory();
+        if(!$this->get('app.functionsHelper')->isConvocatoryValid($current_convocatory)) {
+            $request->getSession()
+                ->getFlashBag()
+                ->add('error', 'Convocatoria antigua (Solo lectura)')
+            ;
+            return $this->redirectToRoute('user_fct');
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($distribution_company);
         $em->flush();
